+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/backupprotectioncontainers_client.go
@@ -46,7 +46,7 @@ func NewBackupProtectionContainersClient(subscriptionID string, credential azcor
 
 // NewListPager - Lists the containers registered to Recovery Services Vault.
 //
-// Generated from API version 2023-02-01
+// Generated from API version 2023-04-01
 //   - vaultName - The name of the recovery services vault.
 //   - resourceGroupName - The name of the resource group where the recovery services vault is present.
 //   - options - BackupProtectionContainersClientListOptions contains the optional parameters for the BackupProtectionContainersClient.NewListPager
@@ -99,7 +99,7 @@ func (client *BackupProtectionContainersClient) listCreateRequest(ctx context.Co
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-02-01")
+	reqQP.Set("api-version", "2023-04-01")
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}