+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/jobs_client.go
@@ -47,7 +47,7 @@ func NewJobsClient(subscriptionID string, credential azcore.TokenCredential, opt
 // Export - Triggers export of jobs specified by filters and returns an OperationID to track.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2023-02-01
+// Generated from API version 2023-04-01
 //   - vaultName - The name of the recovery services vault.
 //   - resourceGroupName - The name of the resource group where the recovery services vault is present.
 //   - options - JobsClientExportOptions contains the optional parameters for the JobsClient.Export method.
@@ -86,7 +86,7 @@ func (client *JobsClient) exportCreateRequest(ctx context.Context, vaultName str
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-02-01")
+	reqQP.Set("api-version", "2023-04-01")
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}