+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/models_serde.go
@@ -2672,6 +2672,7 @@ func (a AzureVMWorkloadProtectableItem) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "friendlyName", a.FriendlyName)
 	populate(objectMap, "isAutoProtectable", a.IsAutoProtectable)
 	populate(objectMap, "isAutoProtected", a.IsAutoProtected)
+	populate(objectMap, "isProtectable", a.IsProtectable)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentUniqueName", a.ParentUniqueName)
 	populate(objectMap, "prebackupvalidation", a.Prebackupvalidation)
@@ -2704,6 +2705,9 @@ func (a *AzureVMWorkloadProtectableItem) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "isAutoProtected":
 			err = unpopulate(val, "IsAutoProtected", &a.IsAutoProtected)
+			delete(rawMsg, key)
+		case "isProtectable":
+			err = unpopulate(val, "IsProtectable", &a.IsProtectable)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -2760,6 +2764,7 @@ func (a AzureVMWorkloadProtectedItem) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "lastBackupStatus", a.LastBackupStatus)
 	populateTimeRFC3339(objectMap, "lastBackupTime", a.LastBackupTime)
 	populateTimeRFC3339(objectMap, "lastRecoveryPoint", a.LastRecoveryPoint)
+	populate(objectMap, "nodesList", a.NodesList)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentType", a.ParentType)
 	populate(objectMap, "policyId", a.PolicyID)
@@ -2836,6 +2841,9 @@ func (a *AzureVMWorkloadProtectedItem) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "lastRecoveryPoint":
 			err = unpopulateTimeRFC3339(val, "LastRecoveryPoint", &a.LastRecoveryPoint)
+			delete(rawMsg, key)
+		case "nodesList":
+			err = unpopulate(val, "NodesList", &a.NodesList)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -3009,6 +3017,7 @@ func (a AzureVMWorkloadSAPAseDatabaseProtectedItem) MarshalJSON() ([]byte, error
 	populate(objectMap, "lastBackupStatus", a.LastBackupStatus)
 	populateTimeRFC3339(objectMap, "lastBackupTime", a.LastBackupTime)
 	populateTimeRFC3339(objectMap, "lastRecoveryPoint", a.LastRecoveryPoint)
+	populate(objectMap, "nodesList", a.NodesList)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentType", a.ParentType)
 	populate(objectMap, "policyId", a.PolicyID)
@@ -3085,6 +3094,9 @@ func (a *AzureVMWorkloadSAPAseDatabaseProtectedItem) UnmarshalJSON(data []byte) 
 			delete(rawMsg, key)
 		case "lastRecoveryPoint":
 			err = unpopulateTimeRFC3339(val, "LastRecoveryPoint", &a.LastRecoveryPoint)
+			delete(rawMsg, key)
+		case "nodesList":
+			err = unpopulate(val, "NodesList", &a.NodesList)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -3206,6 +3218,7 @@ func (a AzureVMWorkloadSAPAseSystemProtectableItem) MarshalJSON() ([]byte, error
 	populate(objectMap, "friendlyName", a.FriendlyName)
 	populate(objectMap, "isAutoProtectable", a.IsAutoProtectable)
 	populate(objectMap, "isAutoProtected", a.IsAutoProtected)
+	populate(objectMap, "isProtectable", a.IsProtectable)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentUniqueName", a.ParentUniqueName)
 	populate(objectMap, "prebackupvalidation", a.Prebackupvalidation)
@@ -3238,6 +3251,9 @@ func (a *AzureVMWorkloadSAPAseSystemProtectableItem) UnmarshalJSON(data []byte) 
 			delete(rawMsg, key)
 		case "isAutoProtected":
 			err = unpopulate(val, "IsAutoProtected", &a.IsAutoProtected)
+			delete(rawMsg, key)
+		case "isProtectable":
+			err = unpopulate(val, "IsProtectable", &a.IsProtectable)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -3344,6 +3360,7 @@ func (a AzureVMWorkloadSAPHanaDBInstance) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "friendlyName", a.FriendlyName)
 	populate(objectMap, "isAutoProtectable", a.IsAutoProtectable)
 	populate(objectMap, "isAutoProtected", a.IsAutoProtected)
+	populate(objectMap, "isProtectable", a.IsProtectable)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentUniqueName", a.ParentUniqueName)
 	populate(objectMap, "prebackupvalidation", a.Prebackupvalidation)
@@ -3376,6 +3393,9 @@ func (a *AzureVMWorkloadSAPHanaDBInstance) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "isAutoProtected":
 			err = unpopulate(val, "IsAutoProtected", &a.IsAutoProtected)
+			delete(rawMsg, key)
+		case "isProtectable":
+			err = unpopulate(val, "IsProtectable", &a.IsProtectable)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -3432,6 +3452,7 @@ func (a AzureVMWorkloadSAPHanaDBInstanceProtectedItem) MarshalJSON() ([]byte, er
 	populate(objectMap, "lastBackupStatus", a.LastBackupStatus)
 	populateTimeRFC3339(objectMap, "lastBackupTime", a.LastBackupTime)
 	populateTimeRFC3339(objectMap, "lastRecoveryPoint", a.LastRecoveryPoint)
+	populate(objectMap, "nodesList", a.NodesList)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentType", a.ParentType)
 	populate(objectMap, "policyId", a.PolicyID)
@@ -3509,6 +3530,9 @@ func (a *AzureVMWorkloadSAPHanaDBInstanceProtectedItem) UnmarshalJSON(data []byt
 		case "lastRecoveryPoint":
 			err = unpopulateTimeRFC3339(val, "LastRecoveryPoint", &a.LastRecoveryPoint)
 			delete(rawMsg, key)
+		case "nodesList":
+			err = unpopulate(val, "NodesList", &a.NodesList)
+			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
 			delete(rawMsg, key)
@@ -3566,6 +3590,7 @@ func (a AzureVMWorkloadSAPHanaDatabaseProtectableItem) MarshalJSON() ([]byte, er
 	populate(objectMap, "friendlyName", a.FriendlyName)
 	populate(objectMap, "isAutoProtectable", a.IsAutoProtectable)
 	populate(objectMap, "isAutoProtected", a.IsAutoProtected)
+	populate(objectMap, "isProtectable", a.IsProtectable)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentUniqueName", a.ParentUniqueName)
 	populate(objectMap, "prebackupvalidation", a.Prebackupvalidation)
@@ -3598,6 +3623,9 @@ func (a *AzureVMWorkloadSAPHanaDatabaseProtectableItem) UnmarshalJSON(data []byt
 			delete(rawMsg, key)
 		case "isAutoProtected":
 			err = unpopulate(val, "IsAutoProtected", &a.IsAutoProtected)
+			delete(rawMsg, key)
+		case "isProtectable":
+			err = unpopulate(val, "IsProtectable", &a.IsProtectable)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -3654,6 +3682,7 @@ func (a AzureVMWorkloadSAPHanaDatabaseProtectedItem) MarshalJSON() ([]byte, erro
 	populate(objectMap, "lastBackupStatus", a.LastBackupStatus)
 	populateTimeRFC3339(objectMap, "lastBackupTime", a.LastBackupTime)
 	populateTimeRFC3339(objectMap, "lastRecoveryPoint", a.LastRecoveryPoint)
+	populate(objectMap, "nodesList", a.NodesList)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentType", a.ParentType)
 	populate(objectMap, "policyId", a.PolicyID)
@@ -3730,6 +3759,9 @@ func (a *AzureVMWorkloadSAPHanaDatabaseProtectedItem) UnmarshalJSON(data []byte)
 			delete(rawMsg, key)
 		case "lastRecoveryPoint":
 			err = unpopulateTimeRFC3339(val, "LastRecoveryPoint", &a.LastRecoveryPoint)
+			delete(rawMsg, key)
+		case "nodesList":
+			err = unpopulate(val, "NodesList", &a.NodesList)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -3844,17 +3876,18 @@ func (a *AzureVMWorkloadSAPHanaDatabaseWorkloadItem) UnmarshalJSON(data []byte) 
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type AzureVMWorkloadSAPHanaHSR.
-func (a AzureVMWorkloadSAPHanaHSR) MarshalJSON() ([]byte, error) {
+// MarshalJSON implements the json.Marshaller interface for type AzureVMWorkloadSAPHanaHSRProtectableItem.
+func (a AzureVMWorkloadSAPHanaHSRProtectableItem) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "backupManagementType", a.BackupManagementType)
 	populate(objectMap, "friendlyName", a.FriendlyName)
 	populate(objectMap, "isAutoProtectable", a.IsAutoProtectable)
 	populate(objectMap, "isAutoProtected", a.IsAutoProtected)
+	populate(objectMap, "isProtectable", a.IsProtectable)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentUniqueName", a.ParentUniqueName)
 	populate(objectMap, "prebackupvalidation", a.Prebackupvalidation)
-	objectMap["protectableItemType"] = "SAPHanaHSR"
+	objectMap["protectableItemType"] = "HanaHSRContainer"
 	populate(objectMap, "protectionState", a.ProtectionState)
 	populate(objectMap, "serverName", a.ServerName)
 	populate(objectMap, "subinquireditemcount", a.Subinquireditemcount)
@@ -3863,8 +3896,8 @@ func (a AzureVMWorkloadSAPHanaHSR) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type AzureVMWorkloadSAPHanaHSR.
-func (a *AzureVMWorkloadSAPHanaHSR) UnmarshalJSON(data []byte) error {
+// UnmarshalJSON implements the json.Unmarshaller interface for type AzureVMWorkloadSAPHanaHSRProtectableItem.
+func (a *AzureVMWorkloadSAPHanaHSRProtectableItem) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
 		return fmt.Errorf("unmarshalling type %T: %v", a, err)
@@ -3883,6 +3916,9 @@ func (a *AzureVMWorkloadSAPHanaHSR) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "isAutoProtected":
 			err = unpopulate(val, "IsAutoProtected", &a.IsAutoProtected)
+			delete(rawMsg, key)
+		case "isProtectable":
+			err = unpopulate(val, "IsProtectable", &a.IsProtectable)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -3926,6 +3962,7 @@ func (a AzureVMWorkloadSAPHanaSystemProtectableItem) MarshalJSON() ([]byte, erro
 	populate(objectMap, "friendlyName", a.FriendlyName)
 	populate(objectMap, "isAutoProtectable", a.IsAutoProtectable)
 	populate(objectMap, "isAutoProtected", a.IsAutoProtected)
+	populate(objectMap, "isProtectable", a.IsProtectable)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentUniqueName", a.ParentUniqueName)
 	populate(objectMap, "prebackupvalidation", a.Prebackupvalidation)
@@ -3958,6 +3995,9 @@ func (a *AzureVMWorkloadSAPHanaSystemProtectableItem) UnmarshalJSON(data []byte)
 			delete(rawMsg, key)
 		case "isAutoProtected":
 			err = unpopulate(val, "IsAutoProtected", &a.IsAutoProtected)
+			delete(rawMsg, key)
+		case "isProtectable":
+			err = unpopulate(val, "IsProtectable", &a.IsProtectable)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -4064,6 +4104,8 @@ func (a AzureVMWorkloadSQLAvailabilityGroupProtectableItem) MarshalJSON() ([]byt
 	populate(objectMap, "friendlyName", a.FriendlyName)
 	populate(objectMap, "isAutoProtectable", a.IsAutoProtectable)
 	populate(objectMap, "isAutoProtected", a.IsAutoProtected)
+	populate(objectMap, "isProtectable", a.IsProtectable)
+	populate(objectMap, "nodesList", a.NodesList)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentUniqueName", a.ParentUniqueName)
 	populate(objectMap, "prebackupvalidation", a.Prebackupvalidation)
@@ -4096,6 +4138,12 @@ func (a *AzureVMWorkloadSQLAvailabilityGroupProtectableItem) UnmarshalJSON(data 
 			delete(rawMsg, key)
 		case "isAutoProtected":
 			err = unpopulate(val, "IsAutoProtected", &a.IsAutoProtected)
+			delete(rawMsg, key)
+		case "isProtectable":
+			err = unpopulate(val, "IsProtectable", &a.IsProtectable)
+			delete(rawMsg, key)
+		case "nodesList":
+			err = unpopulate(val, "NodesList", &a.NodesList)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -4139,6 +4187,7 @@ func (a AzureVMWorkloadSQLDatabaseProtectableItem) MarshalJSON() ([]byte, error)
 	populate(objectMap, "friendlyName", a.FriendlyName)
 	populate(objectMap, "isAutoProtectable", a.IsAutoProtectable)
 	populate(objectMap, "isAutoProtected", a.IsAutoProtected)
+	populate(objectMap, "isProtectable", a.IsProtectable)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentUniqueName", a.ParentUniqueName)
 	populate(objectMap, "prebackupvalidation", a.Prebackupvalidation)
@@ -4171,6 +4220,9 @@ func (a *AzureVMWorkloadSQLDatabaseProtectableItem) UnmarshalJSON(data []byte) e
 			delete(rawMsg, key)
 		case "isAutoProtected":
 			err = unpopulate(val, "IsAutoProtected", &a.IsAutoProtected)
+			delete(rawMsg, key)
+		case "isProtectable":
+			err = unpopulate(val, "IsProtectable", &a.IsProtectable)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -4227,6 +4279,7 @@ func (a AzureVMWorkloadSQLDatabaseProtectedItem) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "lastBackupStatus", a.LastBackupStatus)
 	populateTimeRFC3339(objectMap, "lastBackupTime", a.LastBackupTime)
 	populateTimeRFC3339(objectMap, "lastRecoveryPoint", a.LastRecoveryPoint)
+	populate(objectMap, "nodesList", a.NodesList)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentType", a.ParentType)
 	populate(objectMap, "policyId", a.PolicyID)
@@ -4303,6 +4356,9 @@ func (a *AzureVMWorkloadSQLDatabaseProtectedItem) UnmarshalJSON(data []byte) err
 			delete(rawMsg, key)
 		case "lastRecoveryPoint":
 			err = unpopulateTimeRFC3339(val, "LastRecoveryPoint", &a.LastRecoveryPoint)
+			delete(rawMsg, key)
+		case "nodesList":
+			err = unpopulate(val, "NodesList", &a.NodesList)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -4424,6 +4480,7 @@ func (a AzureVMWorkloadSQLInstanceProtectableItem) MarshalJSON() ([]byte, error)
 	populate(objectMap, "friendlyName", a.FriendlyName)
 	populate(objectMap, "isAutoProtectable", a.IsAutoProtectable)
 	populate(objectMap, "isAutoProtected", a.IsAutoProtected)
+	populate(objectMap, "isProtectable", a.IsProtectable)
 	populate(objectMap, "parentName", a.ParentName)
 	populate(objectMap, "parentUniqueName", a.ParentUniqueName)
 	populate(objectMap, "prebackupvalidation", a.Prebackupvalidation)
@@ -4456,6 +4513,9 @@ func (a *AzureVMWorkloadSQLInstanceProtectableItem) UnmarshalJSON(data []byte) e
 			delete(rawMsg, key)
 		case "isAutoProtected":
 			err = unpopulate(val, "IsAutoProtected", &a.IsAutoProtected)
+			delete(rawMsg, key)
+		case "isProtectable":
+			err = unpopulate(val, "IsProtectable", &a.IsProtectable)
 			delete(rawMsg, key)
 		case "parentName":
 			err = unpopulate(val, "ParentName", &a.ParentName)
@@ -7864,6 +7924,7 @@ func (d DistributedNodesInfo) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "errorDetail", d.ErrorDetail)
 	populate(objectMap, "nodeName", d.NodeName)
+	populate(objectMap, "sourceResourceId", d.SourceResourceID)
 	populate(objectMap, "status", d.Status)
 	return json.Marshal(objectMap)
 }
@@ -7882,6 +7943,9 @@ func (d *DistributedNodesInfo) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "nodeName":
 			err = unpopulate(val, "NodeName", &d.NodeName)
+			delete(rawMsg, key)
+		case "sourceResourceId":
+			err = unpopulate(val, "SourceResourceID", &d.SourceResourceID)
 			delete(rawMsg, key)
 		case "status":
 			err = unpopulate(val, "Status", &d.Status)
@@ -11422,6 +11486,7 @@ func (p *PrivateEndpoint) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaller interface for type PrivateEndpointConnection.
 func (p PrivateEndpointConnection) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
+	populate(objectMap, "groupIds", p.GroupIDs)
 	populate(objectMap, "privateEndpoint", p.PrivateEndpoint)
 	populate(objectMap, "privateLinkServiceConnectionState", p.PrivateLinkServiceConnectionState)
 	populate(objectMap, "provisioningState", p.ProvisioningState)
@@ -11437,6 +11502,9 @@ func (p *PrivateEndpointConnection) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
+		case "groupIds":
+			err = unpopulate(val, "GroupIDs", &p.GroupIDs)
+			delete(rawMsg, key)
 		case "privateEndpoint":
 			err = unpopulate(val, "PrivateEndpoint", &p.PrivateEndpoint)
 			delete(rawMsg, key)
@@ -11508,7 +11576,7 @@ func (p *PrivateEndpointConnectionResource) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaller interface for type PrivateLinkServiceConnectionState.
 func (p PrivateLinkServiceConnectionState) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
-	populate(objectMap, "actionRequired", p.ActionRequired)
+	populate(objectMap, "actionsRequired", p.ActionsRequired)
 	populate(objectMap, "description", p.Description)
 	populate(objectMap, "status", p.Status)
 	return json.Marshal(objectMap)
@@ -11523,8 +11591,8 @@ func (p *PrivateLinkServiceConnectionState) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
-		case "actionRequired":
-			err = unpopulate(val, "ActionRequired", &p.ActionRequired)
+		case "actionsRequired":
+			err = unpopulate(val, "ActionsRequired", &p.ActionsRequired)
 			delete(rawMsg, key)
 		case "description":
 			err = unpopulate(val, "Description", &p.Description)