+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/protectablecontainers_client.go
@@ -46,7 +46,7 @@ func NewProtectableContainersClient(subscriptionID string, credential azcore.Tok
 
 // NewListPager - Lists the containers that can be registered to Recovery Services Vault.
 //
-// Generated from API version 2023-02-01
+// Generated from API version 2023-04-01
 //   - vaultName - The name of the recovery services vault.
 //   - resourceGroupName - The name of the resource group where the recovery services vault is present.
 //   - options - ProtectableContainersClientListOptions contains the optional parameters for the ProtectableContainersClient.NewListPager
@@ -103,7 +103,7 @@ func (client *ProtectableContainersClient) listCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-02-01")
+	reqQP.Set("api-version", "2023-04-01")
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}